refactor: Port `EagerSeriesNamespace`
